Reorganise specs.
