+++ test/path.spec.js
@@ -1,7 +1,6 @@
 var mongoose = require('mongoose');
-var plugin = require('../lib/attachments');
 var User = mongoose.model('User');
-var checksum = require('checksum');
+var plugin = require('../lib/attachments');
 
 describe('path', function(){
 
@@ -15,23 +14,6 @@ describe('path', function(){
         expect(user.profile.original.defaultUrl).to.include('tmp/profile/' + user.id + '-original.png');
         done();
       });
-    });
-  });
-
-});
-
-describe('resampling', function() {
-
-  it('creates a separate image', function(done) {
-    var user = new User({});
-    var path = { path: process.cwd() + '/test/fixtures/mongodb.png' };
-    user.attach('avatar', path, function(err) {
-      checksum.file(user.avatar.thumbnail.path, function(err, generated) {
-        checksum.file(process.cwd() + '/test/fixtures/mongodb-thumbnail-expected.png', function(err, expected) {
-          expect(generated).to.equal(expected);
-        });
-      });
-      done();
     });
   });
